WIP, project lightly restructured into pages and content page outlined
