+++ package.json
@@ -12,7 +12,8 @@
     "axios": "^0.20.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-scripts": "3.4.3",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^3.4.4",
     "react-scroll": "^1.8.1"
   },
   "scripts": {
