+++ src/pages/LandingPage/About.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Banner from "./Banner";
+import Banner from "../../components/Banner";
 import { Element } from "react-scroll";
 
 function About({ id }) {
